+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 == Changelog ==
 
+v3.3.1 - 2017/01/??
+-------------------
+
++ Added `author` support to `llms_membership` post type
++ Added a way to remove all LifterLMS-generated data during plugin uninstallation.
++ Removed references to `LLMS_Activate` class which was removed back in 2.0.
++ Changed include method to session related classes for better handling via phpunit
++ Refactored some of the `LLMS_Install` class for reliability and test coverage
+  + Changed order of table and option creation during installation. Prevents a database warning from being thrown during installation.
+  + Added funciton for retrieving default difficulty categories added during installation
+  + Added function for removing default categories added during installation
++ `llms_are_terms_and_conditions_required()` ensure the page id used in this function is an absint
++ Fix load method of LifterLMS Post Model classes
++ `llms_get_post()` will now work with any LifterLMS Post Model post types
+
+
 v3.3.0 - 2017/01/23
 -------------------